nicer colors. nicer donate page
