+++ src/components/Sidebar.module.css
@@ -8,9 +8,9 @@
   display: flex;
   flex-direction: column;
   position: fixed;
-  top: 150px;
+  top: 140px;
   left: 0;
-  height: calc(100vh - 150px);
+  height: calc(100vh - 140px);
   transition: top 0.2s ease-in-out, width 0.4s ease;
   background: var(--lightgray);
   width: var(--margin);
