+++ src/components/character/Donate.module.css
@@ -21,6 +21,7 @@
   }
   & p {
     line-height: 1.2;
+    padding-bottom: 0.5rem;
   }
 }
 .root:hover {
@@ -31,6 +32,7 @@
 }
 .compact .text {
   max-height: 0;
+  padding-bottom: 0;
 }
 .compact:hover .text {
   max-height: 100px;
