+++ src/styles/variables.css
@@ -12,10 +12,9 @@
   --processing-blue-light: #82afff;
   --processing-blue-mid: #0564ff;
   --processing-blue-deep: #1e32aa;
-  --processing-blue-dark: #1e2878;
-  /* --processing-blue-dark: #1e32aa; */
+  --processing-blue-dark: #0f195a;
 
-  --topbar-background: #0a1428;
+  --topbar-background: #02163e;
 
   --processing-blue-gradient: #508bff2e;
